+++ verificationmanagement_test.go
@@ -1,0 +1,138 @@
+// File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
+
+package prelude_test
+
+import (
+	"context"
+	"errors"
+	"os"
+	"testing"
+
+	"github.com/prelude-so/go-sdk"
+	"github.com/prelude-so/go-sdk/internal/testutil"
+	"github.com/prelude-so/go-sdk/option"
+)
+
+func TestVerificationManagementDeletePhoneNumber(t *testing.T) {
+	baseURL := "http://localhost:4010"
+	if envURL, ok := os.LookupEnv("TEST_API_BASE_URL"); ok {
+		baseURL = envURL
+	}
+	if !testutil.CheckTestServer(t, baseURL) {
+		return
+	}
+	client := prelude.NewClient(
+		option.WithBaseURL(baseURL),
+		option.WithAPIToken("My API Token"),
+	)
+	_, err := client.VerificationManagement.DeletePhoneNumber(
+		context.TODO(),
+		prelude.VerificationManagementDeletePhoneNumberParamsActionAllow,
+		prelude.VerificationManagementDeletePhoneNumberParams{
+			PhoneNumber: prelude.F("+30123456789"),
+		},
+	)
+	if err != nil {
+		var apierr *prelude.Error
+		if errors.As(err, &apierr) {
+			t.Log(string(apierr.DumpRequest(true)))
+		}
+		t.Fatalf("err should be nil: %s", err.Error())
+	}
+}
+
+func TestVerificationManagementListPhoneNumbers(t *testing.T) {
+	baseURL := "http://localhost:4010"
+	if envURL, ok := os.LookupEnv("TEST_API_BASE_URL"); ok {
+		baseURL = envURL
+	}
+	if !testutil.CheckTestServer(t, baseURL) {
+		return
+	}
+	client := prelude.NewClient(
+		option.WithBaseURL(baseURL),
+		option.WithAPIToken("My API Token"),
+	)
+	_, err := client.VerificationManagement.ListPhoneNumbers(context.TODO(), prelude.VerificationManagementListPhoneNumbersParamsActionAllow)
+	if err != nil {
+		var apierr *prelude.Error
+		if errors.As(err, &apierr) {
+			t.Log(string(apierr.DumpRequest(true)))
+		}
+		t.Fatalf("err should be nil: %s", err.Error())
+	}
+}
+
+func TestVerificationManagementListSenderIDs(t *testing.T) {
+	baseURL := "http://localhost:4010"
+	if envURL, ok := os.LookupEnv("TEST_API_BASE_URL"); ok {
+		baseURL = envURL
+	}
+	if !testutil.CheckTestServer(t, baseURL) {
+		return
+	}
+	client := prelude.NewClient(
+		option.WithBaseURL(baseURL),
+		option.WithAPIToken("My API Token"),
+	)
+	_, err := client.VerificationManagement.ListSenderIDs(context.TODO())
+	if err != nil {
+		var apierr *prelude.Error
+		if errors.As(err, &apierr) {
+			t.Log(string(apierr.DumpRequest(true)))
+		}
+		t.Fatalf("err should be nil: %s", err.Error())
+	}
+}
+
+func TestVerificationManagementSetPhoneNumber(t *testing.T) {
+	baseURL := "http://localhost:4010"
+	if envURL, ok := os.LookupEnv("TEST_API_BASE_URL"); ok {
+		baseURL = envURL
+	}
+	if !testutil.CheckTestServer(t, baseURL) {
+		return
+	}
+	client := prelude.NewClient(
+		option.WithBaseURL(baseURL),
+		option.WithAPIToken("My API Token"),
+	)
+	_, err := client.VerificationManagement.SetPhoneNumber(
+		context.TODO(),
+		prelude.VerificationManagementSetPhoneNumberParamsActionAllow,
+		prelude.VerificationManagementSetPhoneNumberParams{
+			PhoneNumber: prelude.F("+30123456789"),
+		},
+	)
+	if err != nil {
+		var apierr *prelude.Error
+		if errors.As(err, &apierr) {
+			t.Log(string(apierr.DumpRequest(true)))
+		}
+		t.Fatalf("err should be nil: %s", err.Error())
+	}
+}
+
+func TestVerificationManagementSubmitSenderID(t *testing.T) {
+	baseURL := "http://localhost:4010"
+	if envURL, ok := os.LookupEnv("TEST_API_BASE_URL"); ok {
+		baseURL = envURL
+	}
+	if !testutil.CheckTestServer(t, baseURL) {
+		return
+	}
+	client := prelude.NewClient(
+		option.WithBaseURL(baseURL),
+		option.WithAPIToken("My API Token"),
+	)
+	_, err := client.VerificationManagement.SubmitSenderID(context.TODO(), prelude.VerificationManagementSubmitSenderIDParams{
+		SenderID: prelude.F("Prelude"),
+	})
+	if err != nil {
+		var apierr *prelude.Error
+		if errors.As(err, &apierr) {
+			t.Log(string(apierr.DumpRequest(true)))
+		}
+		t.Fatalf("err should be nil: %s", err.Error())
+	}
+}